hacker_game/requirements.txt
PySide6>=6.5.0
pygame>=2.5.0
numpy>=1.24.0
Pillow>=10.0.0